Fix build failures on 32-bit and optimized targets
size_t is a 32-bit unsigned int on armv7, so the Content-Length values
in rest.c and the send length in fcgi.c - all strlen() results printed
with %ld or %lu - failed -Werror=format there. Print them with %zu
instead. The content-id in yang-library.c has the same problem in
reverse: time_t is a long on that target while PRIx64 expects a long
long, so cast it to uint64_t.

Building for ppc64_e6500mc separately failed -Werror=maybe-uninitialized
on "rc" in rest_api_post(), reported against the copy of
log_post_put_patch() inlined at the exit label. Every path through the
function does assign rc before reaching there, but initialise it at its
declaration to the success value anyway, matching the other
rest_api_*() handlers.

+++ yang-library.c
@@ -198,7 +198,7 @@ yang_library_create (sch_instance *schema)
 
     /* Each time this routine is run the content-id will be set to a unique id based
      * on the clock */
-    snprintf (set_id, sizeof (set_id), "%" PRIx64 "", now);
+    snprintf (set_id, sizeof (set_id), "%" PRIx64 "", (uint64_t) now);
     apteryx_set (YANG_LIBRARY_CONTENT_ID, set_id);
 }
